+++ lib/auth/registry_credentials.go
@@ -3,6 +3,7 @@ package auth
 type RegistryCredentials interface {
 	User() string
 	Password() string
+	IdentityToken() string
 	SetPassword(password string)
 	SetUser(user string)
 }